Refactor how completions are organized.
