+++ qutebrowser/completion/models/configmodel.py
@@ -17,13 +17,12 @@
 # You should have received a copy of the GNU General Public License
 # along with qutebrowser.  If not, see <http://www.gnu.org/licenses/>.
 
-"""CompletionModels for different usages."""
+"""CompletionModels for the config."""
 
 from PyQt5.QtCore import pyqtSlot, Qt
 
 from qutebrowser.config import config, configdata
 from qutebrowser.utils import log, qtutils, objreg
-from qutebrowser.commands import cmdutils
 from qutebrowser.completion.models import base
 
 
@@ -148,168 +147,3 @@ class SettingValueCompletionModel(base.BaseCompletionModel):
         if not ok:
             raise ValueError("Setting data failed! (section: {}, option: {}, "
                              "value: {})".format(section, option, value))
-
-
-class CommandCompletionModel(base.BaseCompletionModel):
-
-    """A CompletionModel filled with all commands and descriptions."""
-
-    # pylint: disable=abstract-method
-
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        assert cmdutils.cmd_dict
-        cmdlist = []
-        for obj in set(cmdutils.cmd_dict.values()):
-            if (obj.hide or (obj.debug and not objreg.get('args').debug) or
-                    obj.deprecated):
-                pass
-            else:
-                cmdlist.append((obj.name, obj.desc))
-        for name, cmd in config.section('aliases').items():
-            cmdlist.append((name, "Alias for '{}'".format(cmd)))
-        cat = self.new_category("Commands")
-        for (name, desc) in sorted(cmdlist):
-            self.new_item(cat, name, desc)
-
-
-class HelpCompletionModel(base.BaseCompletionModel):
-
-    """A CompletionModel filled with help topics."""
-
-    # pylint: disable=abstract-method
-
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self._init_commands()
-        self._init_settings()
-
-    def _init_commands(self):
-        """Fill completion with :command entries."""
-        assert cmdutils.cmd_dict
-        cmdlist = []
-        for obj in set(cmdutils.cmd_dict.values()):
-            if (obj.hide or (obj.debug and not objreg.get('args').debug) or
-                    obj.deprecated):
-                pass
-            else:
-                cmdlist.append((':' + obj.name, obj.desc))
-        cat = self.new_category("Commands")
-        for (name, desc) in sorted(cmdlist):
-            self.new_item(cat, name, desc)
-
-    def _init_settings(self):
-        """Fill completion with section->option entries."""
-        cat = self.new_category("Settings")
-        for sectname, sectdata in configdata.DATA.items():
-            for optname in sectdata.keys():
-                try:
-                    desc = sectdata.descriptions[optname]
-                except (KeyError, AttributeError):
-                    # Some stuff (especially ValueList items) don't have a
-                    # description.
-                    desc = ""
-                else:
-                    desc = desc.splitlines()[0]
-                name = '{}->{}'.format(sectname, optname)
-                self.new_item(cat, name, desc)
-
-
-class UrlCompletionModel(base.BaseCompletionModel):
-
-    """A model which combines quickmarks and web history URLs.
-
-    Used for the `open` command."""
-
-    # pylint: disable=abstract-method
-
-    def __init__(self, parent=None):
-        super().__init__(parent)
-
-        self._quickcat = self.new_category("Quickmarks")
-        self._histcat = self.new_category("History")
-        self._histstore = objreg.get('web-history')
-
-        WebHistoryCompletionModel.fill_model(self, cat=self._histcat)
-        QuickmarkCompletionModel.fill_model(self, cat=self._quickcat)
-
-        self._histstore.item_added.connect(
-            lambda e: WebHistoryCompletionModel.history_changed(
-                self, e, self._histcat))
-
-
-class WebHistoryCompletionModel(base.BaseCompletionModel):
-
-    """A CompletionModel filled with global browsing history."""
-
-    # pylint: disable=abstract-method
-
-    def __init__(self, parent=None):
-        super().__init__(parent)
-
-        self._histcat = self.new_category("History")
-        self._histstore = objreg.get('web-history')
-
-        self.fill_model(self, self._histcat, self._histstore)
-
-        self._histstore.item_added.connect(
-            lambda e: self.history_changed(e, self._histcat))
-
-    @staticmethod
-    def fill_model(model, cat=None, histstore=None):
-        """Fill the given model/category with the given history."""
-        if not histstore:
-            histstore = objreg.get('web-history')
-        if not cat:
-            cat = model.new_category("History")
-
-        for entry in histstore:
-            atime = int(entry.atime)
-            model.new_item(cat, entry.url, "", str(atime), sort=atime)
-
-    def history_changed(self, entry, cat):
-        """Slot called when a new history item was added."""
-        if entry.url:
-            atime = int(entry.atime)
-            self.new_item(cat, entry.url, "", str(atime), sort=atime)
-
-
-class QuickmarkCompletionModel(base.BaseCompletionModel):
-
-    """A CompletionModel filled with all quickmarks."""
-
-    # pylint: disable=abstract-method
-
-    def __init__(self, match_field='url', parent=None):
-        super().__init__(parent)
-        self.fill_model(self, match_field)
-
-    @staticmethod
-    def fill_model(model, match_field='url', cat=None):
-        """Fill the given model/category with quickmarks."""
-        if not cat:
-            cat = model.new_category("Quickmarks")
-        quickmarks = objreg.get('quickmark-manager').marks.items()
-
-        if match_field == 'url':
-            for qm_name, qm_url in quickmarks:
-                model.new_item(cat, qm_url, qm_name)
-        elif match_field == 'name':
-            for qm_name, qm_url in quickmarks:
-                model.new_item(cat, qm_name, qm_url)
-        else:
-            raise ValueError("Invalid value '{}' for match_field!".format(
-                match_field))
-
-
-class SessionCompletionModel(base.BaseCompletionModel):
-
-    """A CompletionModel filled with session names."""
-
-    # pylint: disable=abstract-method
-
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        cat = self.new_category("Sessions")
-        for name in objreg.get('session-manager').list_sessions():
-            self.new_item(cat, name)
